Add 2nd option for bin widths in sfc station rank plots
New bins are wider, which helps fix the problem in the original figure where there are very few values in the 0.5 bin (this occurs b/c if there are 29 obs instead of the full 30, none of the decimal ranks falls into the 0.5 bin b/c it is so narrow). These new bins, however, have unequal bin widths, with the first and last bins being narrower than the rest, resulting in what looks like fewer values in these bins.  did this on purpose b/c I wanted to clearly see how many NR values lied outside of the observed climatology.